pyproject.toml
[tool.poetry]
name = "openai-sdk-agent"
version = "0.1.0"
description = "An automated sales agent system using OpenAI SDK for cold email generation"
authors = ["Robin Ochieng <your.email@example.com>"]
readme = "README.md"
packages = [{include = "openai_sdk_agent.py"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.107.0"
openai-agents = "^0.3.3"
python-dotenv = "^1.0.0"
sendgrid = "^6.11.0"
asyncio = "^3.4.3"
gradio = "^5.0.0"
pydantic = "^2.10.0"
fastapi = "^0.119.0"
uvicorn = "^0.37.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
